api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
azure-functions-durable>=1.2.10
azure-storage-blob
openai
matplotlib
azure-cognitiveservices-vision-customvision
msrest
python-dotenv
ipykernel
pillow
streamlit
azure-identity
